engine logic exported to Engine class, minor improvements

+++ gun.h
@@ -28,5 +28,5 @@ private:
 
     int attack;
     unsigned int speed;
-    float range;
+    unsigned int range;
 };
